refactor: make hyperkzg a directory module
With hyperkzg becoming more complicated, and some components being gated
by different feature flags, the file-module holding it has become too
large and confusing. This change makes hyperkzg a directory module, and
moves every individual component to submodules.
